moved test_boxy.py to io instead of preprocessing, same as test_nirx.py
